+++ Options/CHANGELOG.lua
@@ -5,6 +5,7 @@ TMW.CHANGELOG_LASTVER="7.4.0"
 TMW.CHANGELOG = [==[
 ## v9.0.6
 * New Condition: Covenant Membership
+* New group sort option: Timer start
 
 ### Bug Fixes
 * Fixed an issue with some types of items in the item suggestion list.