+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## v9.0.6
 * New Condition: Covenant Membership
+* New group sort option: Timer start
 
 ### Bug Fixes
 * Fixed an issue with some types of items in the item suggestion list.